backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
environs==14.5.0
fastapi==0.129.0
fastapi_cors==0.0.6
h11==0.16.0
idna==3.11
marshmallow==4.2.2
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
